Hook for form state and validation + fix ProtectedRoute
ProtectedRoute was only checking if there is token in the localStorage,
and displaying the component even if the token was expired.
I Added a call to new endpoint /token to check current token validity
and display component if it is valid.

Changed managing state of the form in the custom hook `useForm`, which
also calls validate with provided custom field validators insted of
predefined validators.